requirements-dev.txt
-r requirements.txt
coverage>=7.4.4
pytest==8.3.4
pytest-django==4.9.0
ruff==0.9.10
bandit==1.8.2
pip-audit==2.8.0
